+++ src/ui/camera.test.ts
@@ -1799,6 +1799,8 @@ describe('#flyTo', () => {
         camera.transform = {
             elevation: 0,
             recalculateZoom: () => true,
+            clone: () => camera.transform,
+            apply: () => {},
             setMinElevationForCurrentTile: (_a) => true,
             setElevation: (e) => { camera.transform.elevation = e; }
         };