+++ src/ui/camera.ts
@@ -1017,7 +1017,9 @@ export abstract class Camera extends Evented {
 
     _finalizeElevation() {
         this._elevationFreeze = false;
-        this.transform.recalculateZoom(this.terrain);
+        const tr = this._getTransformForUpdate();
+        tr.recalculateZoom(this.terrain);
+        this._applyUpdatedTransform(tr);
     }
 
     /**
@@ -1049,10 +1051,11 @@ export abstract class Camera extends Evented {
      *
      * @param tr - The transform to check.
      */
-    _elevateCameraIfInsideTerrain(tr: ITransform) : { pitch?: number; zoom?: number } {
+    _elevateCameraIfInsideTerrain(tr: ITransform): { pitch?: number; zoom?: number } {
         const cameraLngLat = tr.screenPointToLocation(tr.getCameraPoint());
         const cameraAltitude = tr.getCameraAltitude();
-        const minAltitude = this.terrain.getElevationForLngLatZoom(cameraLngLat, tr.zoom);
+        const surfacePadding = Math.min(500, 20 * (25 - tr.zoom));
+        const minAltitude = this.terrain.getElevationForLngLatZoom(cameraLngLat, tr.zoom) + surfacePadding;
         if (cameraAltitude < minAltitude) {
             const newCamera = this.calculateCameraOptionsFromTo(
                 cameraLngLat, minAltitude, tr.center, tr.elevation);